requirements.txt
boto3
requests
psycopg2-binary
python-dotenv
json-repair>=0.30.0
